change: separate reads into diff elements at parse
When samplesheet is parsed, emit samples as lists of
[ metadata, [reads1], [reads2] ].
If reads2 does not exist, supply an empty file as a placeholder.
All logic for single- or paired-end reads MUST depend on the
metadata.readType attribute.
The goal here is to simplify grouping reads and create a
common interface to processes so that they can take either grouped or
individual reads samples.